posix: isolate non-portable pthread APIs under CONFIG_PTHREAD_NP
Fixes #98536.

This isolates non-portable pthread extensions (e.g. pthread_setname_np) behind the new CONFIG_PTHREAD_NP flag, ensuring better strict POSIX compliance. Also adds a new test suite under tests/posix/pthread_np to verify these non-portable APIs independently.
