+++ samples/subsys/portability/posix/philosophers/prj.conf
@@ -11,3 +11,4 @@ CONFIG_MAX_PTHREAD_MUTEX_COUNT=6
 
 #Enable thread awareness for debugging tools supporting it
 CONFIG_DEBUG_THREAD_INFO=y
+CONFIG_PTHREAD_NP=y
