+++ subsys/portability/posix/options/pthread.c
@@ -1142,8 +1142,11 @@ static int pthread_timedjoin_internal(pthread_t pthread, void **status, k_timeou
 	return 0;
 }
 
+#if defined(CONFIG_PTHREAD_NP) || defined(__DOXYGEN__)
 /**
  * @brief Await a thread termination with timeout.
+ *
+ * @kconfig_dep{CONFIG_PTHREAD_NP}
  *
  * Non-portable GNU extension of IEEE 1003.1
  */
@@ -1161,12 +1164,15 @@ int pthread_timedjoin_np(pthread_t pthread, void **status, const struct timespec
 /**
  * @brief Check a thread for termination.
  *
+ * @kconfig_dep{CONFIG_PTHREAD_NP}
+ *
  * Non-portable GNU extension of IEEE 1003.1
  */
 int pthread_tryjoin_np(pthread_t pthread, void **status)
 {
 	return pthread_timedjoin_internal(pthread, status, K_NO_WAIT);
 }
+#endif /* CONFIG_PTHREAD_NP */
 
 /**
  * @brief Await a thread termination.
@@ -1435,6 +1441,7 @@ int pthread_attr_destroy(pthread_attr_t *_attr)
 	return 0;
 }
 
+#if defined(CONFIG_PTHREAD_NP) || defined(__DOXYGEN__)
 int pthread_setname_np(pthread_t thread, const char *name)
 {
 #ifdef CONFIG_THREAD_NAME
@@ -1483,6 +1490,7 @@ int pthread_getname_np(pthread_t thread, char *name, size_t len)
 	return 0;
 #endif
 }
+#endif /* CONFIG_PTHREAD_NP */
 
 int pthread_atfork(void (*prepare)(void), void (*parent)(void), void (*child)(void))
 {
