+++ tests/posix/pthread_np/prj.conf
@@ -1,0 +1,8 @@
+# SPDX-FileCopyrightText: Copyright The Zephyr Project Contributors
+# SPDX-License-Identifier: Apache-2.0
+
+CONFIG_ZTEST=y
+CONFIG_ZTEST_NEW_API=y
+CONFIG_POSIX_API=y
+CONFIG_PTHREAD_NP=y
+CONFIG_THREAD_NAME=y
