+++ tests/posix/pthread_np/testcase.yaml
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: Copyright The Zephyr Project Contributors
+# SPDX-License-Identifier: Apache-2.0
+
+tests:
+  portability.posix.pthread_np:
+    tags: posix
+    depends_on: posix
